+++ src/main/java/io/hhplus/tdd/point/PointValidator.java
@@ -1,4 +1,4 @@
-package io.hhplus.tdd.point.service;
+package io.hhplus.tdd.point;
 
 import org.springframework.stereotype.Component;
